+++ src/call.rs
@@ -33,8 +33,7 @@ use ton_client::processing::{
     wait_for_transaction,
     send_message,
 };
-use ton_client::tvm::{run_tvm, run_get, ParamsOfRunTvm, ParamsOfRunGet};
-use ton_client::error::ClientError;
+use ton_client::tvm::{run_tvm, run_get, ParamsOfRunTvm, ParamsOfRunGet, run_executor, ParamsOfRunExecutor, AccountForExecutor};
 
 pub struct EncodedMessage {
     pub message_id: String,
@@ -245,6 +244,33 @@ async fn query_account_boc(ton: TonClient, addr: &str) -> Result<String, String>
     Ok(boc.unwrap().to_owned())
 }
 
+pub async fn emulate_localy(
+    ton: TonClient,
+    addr: &str,
+    msg: String,
+) -> Result<(), String> {
+    let state_boc = query_account_boc(ton.clone(), addr).await?;
+
+    let res = run_executor(
+        ton.clone(),
+        ParamsOfRunExecutor {
+            message: msg.clone(),
+            account: AccountForExecutor::Account {
+                boc: state_boc,
+                unlimited_balance: None,
+            },
+            ..Default::default()
+        },
+    )
+    .await;
+
+    if res.is_err() {
+        return Err(format!("Local execution has failed:\n{:#}", res.err().unwrap()));
+    }
+    println!("Local run succeeded. Executing onchain.");
+    Ok(())
+}
+
 async fn send_message_and_wait(
     ton: TonClient,
     addr: &str,
@@ -252,7 +278,6 @@ async fn send_message_and_wait(
     msg: String,
     local: bool,
     conf: Config,
-    mut error_handler: impl FnMut(ClientError),
 ) -> Result<serde_json::Value, String> {
     if local {
         if !conf.is_json {
@@ -292,7 +317,6 @@ async fn send_message_and_wait(
         ).await;
         if result.is_err() {
             let err = result.err().unwrap();
-            error_handler(err.clone());
             return Err(format!("{:#}", err));
         }
 
@@ -309,7 +333,6 @@ async fn send_message_and_wait(
         ).await;
         if result.is_err() {
             let err = result.err().unwrap();
-            error_handler(err.clone());
             return Err(format!("{:#}", err));
         }
         Ok(result.unwrap().decoded.and_then(|d| d.output).unwrap_or(json!({})))
@@ -353,42 +376,17 @@ pub async fn call_contract_with_result(
         if !conf.is_json {
             print_encoded_message(&msg);
         }
-        
-        let mut retry: bool = true;
-        let error_handler = |err: ClientError| {
-            // obtaining error code
-            let code = err.code.clone();
-            // but if it was simulated locally and local exit code is not zero,
-            // we ignore previous exit code because it means we shouldn't make a retry.
-            if !err.data["exit_code"].is_null() {
-                if err.data["exit_code"].as_i64().unwrap() != 0 {
-                    retry = false;
-                }
-            }
-            // There is also another way how SDK can print local run results.
-            let local_error = err.data["local_error"]["data"]["exit_code"].clone();
-            if !local_error.is_null() {
-                if local_error.as_i64().unwrap() != 0 {
-                    retry = false;
-                }
-            }
-            // if error code was 4XX then don't perform a retry.
-            let code = (code / 100) as u32 % 10;
-            if  code == 4 {
-                retry = false;
-            }
-        };
-        let result = send_message_and_wait(ton.clone(), addr, abi.clone(), msg.message, local, conf.clone(), error_handler).await;
+
+        if !local && !conf.no_local_run {
+            emulate_localy(ton.clone(), addr, msg.message.clone()).await?;
+        }
+        let result = send_message_and_wait(ton.clone(), addr, abi.clone(), msg.message, local, conf.clone()).await;
         
         if result.is_ok() {
             return result;
         }
         let err = result.err().unwrap();
         println!("{}", err);
-
-        if !retry {
-            break;
-        }
 
         if attempts != 0 {
             println!("\nRetry #{}.\n", total_attempts - attempts);
@@ -498,7 +496,7 @@ pub async fn call_contract_with_msg(conf: Config, str_msg: String, abi: String) 
     println!("{}", params.1);
     println!("Processing... ");
 
-    let result = send_message_and_wait(ton, &msg.address, abi, msg.message, false, conf, |_| {}).await?;
+    let result = send_message_and_wait(ton, &msg.address, abi, msg.message, false, conf).await?;
 
     println!("Succeded.");
     if !result.is_null() {