+++ src/config.rs
@@ -67,6 +67,8 @@ pub struct Config {
     pub no_answer: bool,
     #[serde(default = "default_true")]
     pub use_delimiters: bool,
+    #[serde(default = "default_false")]
+    pub no_local_run: bool,
 }
 
 impl Config {
@@ -85,6 +87,7 @@ impl Config {
             lifetime: default_lifetime(),
             no_answer: default_false(),
             use_delimiters: default_true(),
+            no_local_run: default_false(),
         }
     }
 
@@ -110,6 +113,7 @@ pub fn clear_config(
     lifetime: bool,
     no_answer: bool,
     use_delimiters: bool,
+    no_local_run: bool,
 ) -> Result<(), String> {
     if url {
         conf.url = default_url();
@@ -147,9 +151,12 @@ pub fn clear_config(
     if use_delimiters {
         conf.use_delimiters = default_true();
     }
+    if no_local_run {
+        conf.no_local_run = default_false();
+    }
 
     if (url || addr || wallet || abi || keys || retries || timeout || wc || depool_fee || lifetime
-        || no_answer || use_delimiters) == false {
+        || no_answer || use_delimiters || no_local_run) == false {
         conf = Config {
             url: default_url(),
             wc: default_wc(),
@@ -164,6 +171,7 @@ pub fn clear_config(
             lifetime: default_lifetime(),
             no_answer: default_false(),
             use_delimiters: default_true(),
+            no_local_run: default_false(),
         };
     }
     let conf_str = serde_json::to_string(&conf)
@@ -189,6 +197,7 @@ pub fn set_config(
     lifetime:  Option<&str>,
     no_answer:  Option<&str>,
     use_delimiters: Option<&str>,
+    no_local_run: Option<&str>,
 ) -> Result<(), String> {
     if let Some(s) = url {
         conf.url = s.to_string();
@@ -235,6 +244,10 @@ pub fn set_config(
     if let Some(use_delimiters) = use_delimiters {
         conf.use_delimiters = use_delimiters.parse::<bool>()
             .map_err(|e| format!(r#"failed to parse "use_delimiters": {}"#, e))?;
+    }
+    if let Some(no_local_run) = no_local_run {
+        conf.no_local_run = no_local_run.parse::<bool>()
+            .map_err(|e| format!(r#"failed to parse "no_local_run": {}"#, e))?;
     }
 
     let conf_str = serde_json::to_string(&conf)